move errors to reduce clutter
